allow passwords to be supplied in hex-encoded format
The current method for accepting passwords in secure mode does not
allow entering arbitrary bytes.  For example, byte value 10 (newline)
will terminate input, and byte value 127 (del) will erase the
last-entered byte.  This makes it impossible to use arbitrary bytes in
keys.  This patch adds an `-x` option which will instruct the
password-reading code to treat it as a hex-encoded string.

This ability is available for the following operations that accept
password input:

- `setAdmin1Pwd`
- `setLockingRange`
- `listLockingRange`
- `enableLockingRange`

This support is not available in the `--setSIDPassword` operation.